+++ KhaozEngine.NetWorld/WorldMetaRecord.cs
@@ -14,8 +14,10 @@ public sealed class WorldMetaRecord
     /// <summary>Record schema version; bump when the shape changes meaningfully.</summary>
     public int Version { get; set; } = 1;
 
-    /// <summary>The next NetId the allocator will hand out (one past the highest ever allocated).</summary>
-    public int NextNetId { get; set; }
+    /// <summary>The next NetId the allocator will hand out (one past the highest ever allocated), as a packed 64-bit
+    /// value for the local node. Widened from <c>int</c> in 10.0.0; JSON is forward-tolerant, so an old 32-bit value
+    /// parses into this <c>long</c> unchanged.</summary>
+    public long NextNetId { get; set; }
 
     /// <summary>Serializes to UTF-8 JSON bytes for the world store.</summary>
     public byte[] Encode() => JsonSerializer.SerializeToUtf8Bytes(this, JsonDefaults.IndentedWrite);
